Fixes apollo-server version testing so that the permutation under test is loaded.
