+++ tests/integration/state-loss.test.js
@@ -8,7 +8,7 @@
 const { executeQuery } = require('../test-client')
 const { setupEnvConfig } = require('../agent-testing')
 
-const { setupApolloServerTests } = require('../apollo-server-setup')
+const { setupApolloServerTests } = require('./apollo-server-setup')
 
 const PluginStateLossTester = require('./plugin-state-loss-tester')
 const stateLossTester = new PluginStateLossTester()
