+++ tests/versioned/apollo-server/segments.test.js
@@ -5,7 +5,7 @@
 
 'use strict'
 
-const { setupApolloServerTests } = require('../../apollo-server-setup')
+const { setupApolloServerTests } = require('./apollo-server-setup')
 const expressSegmentsDefaultTests = require('../express-segments-default-tests')
 const expressSegmentsScalarTests = require('../express-segments-scalar-tests')
